# backend/nail-resv/openapi-complete.yaml
...
openapi: 3.0.3
info:
  title: 美甲預約系統完整 API
  description: |
    完整的美甲預約系統 API，包含用戶註冊登入、美甲師管理、預約系統、圖片標註和搜尋功能。
    
    ## 主要功能
    - 🔐 用戶註冊登入系統
    - 👤 顧客和美甲師資料管理
    - 📅 預約和時段管理
    - 🎨 美甲圖片 AI 標註
    - 🔍 智能搜尋和篩選
    - ⏰ 美甲師營業時間設定
  version: 1.0.0
  contact:
    name: 美甲預約系統 API 支援
    email: support@nailit.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: 開發環境
  - url: https://api.nailit.com/api
    description: 生產環境
paths:
...
  /customers/{id}:
    get:
      tags: [顧客管理]
      summary: 獲取顧客資料
      description: 根據用戶ID獲取顧客詳細資料
      operationId: getCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 顧客的用戶ID
      responses:
        '200':
          description: 獲取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: 顧客不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [顧客管理]
      summary: 更新顧客資料
      description: 更新顧客的基本資料
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 顧客的用戶ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCustomerResponse'
  /customers/{id}/appointments:
    get:
      tags: [顧客管理]
      summary: 獲取顧客預約記錄
      description: 獲取指定顧客的所有預約記錄
      operationId: getCustomerAppointments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 顧客的用戶ID
      responses:
        '200':
          description: 獲取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAppointmentsResponse'

  # ========== 美甲師管理 API ==========
...
  /reservations/customer/{customerId}:
    get:
      tags: [預約管理]
      summary: 獲取顧客的預約列表
      description: 獲取指定顧客的所有預約記錄
      operationId: getCustomerReservations
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
          description: 顧客的用戶ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
          description: 篩選預約狀態
      responses:
        '200':
          description: 獲取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerReservationsResponse'
...
components:
  schemas:
...
    CustomerResponse:
      type: object
      properties:
        success:
          type: boolean
        customer:
          $ref: '#/components/schemas/Customer'
    Customer:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        created_at:
          type: string
          format: date-time
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
    UpdateCustomerResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        customer:
          type: object
    CustomerAppointmentsResponse:
      type: object
      properties:
        success:
          type: boolean
        appointments:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'

    # ========== 美甲師管理 Schema ==========
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        customerName:
          type: string
        artistName:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        note:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CustomerReservationsResponse:
      type: object
      properties:
        success:
          type: boolean
        customerId:
          type: string
        customerName:
          type: string
        appointments:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        total:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: 錯誤訊息
          example: "請求參數錯誤"
        detail:
          type: string
          description: 詳細錯誤資訊
          example: "Missing required field: customerId"
...